+++ abstracto-application/abstracto-modules/starboard/starboard-int/src/main/java/dev/sheldan/abstracto/starboard/model/template/GuildStarStatsModel.java
@@ -15,5 +15,5 @@ public class GuildStarStatsModel {
     private List<StarStatsUser> starGiver;
     private Integer totalStars;
     private List<String> badgeEmotes;
-    private Integer starredMessages;
+    private Long starredMessages;
 }
